requirements.txt
pypdf2~=2.11.2
fitz~=0.0.1.dev2
pyzbar~=0.1.9
Pillow~=9.3.0
PyMuPDF==1.21.0
